fixed persistence
