+++ data/customers.json
@@ -1,1 +1,1 @@
-[{"username":"testing","bankAccounts":[{"balance":900.0,"accountNumber":1,"transactionHistory":[{"type":"Deposit","amount":999.0,"accountNumber":1},{"type":"Withdrawal","amount":9.0,"accountNumber":1},{"type":"Transfer Out","amount":90.0,"accountNumber":2}]},{"balance":90.0,"accountNumber":2,"transactionHistory":[{"type":"Transfer In","amount":90.0,"accountNumber":1}]}]},{"username":"newmax","bankAccounts":[],"mortgage":{"loanAmount":100.0,"annualRate":0.05,"termYears":10.0,"remainingAmount":10.0}},{"username":"brad","bankAccounts":[],"mortgage":{"loanAmount":100.0,"annualRate":0.06,"termYears":67.0,"remainingAmount":100.0}},{"username":"max","bankAccounts":[{"balance":0.0,"accountNumber":3,"transactionHistory":[{"type":"Deposit","amount":300.0,"accountNumber":3},{"type":"Withdrawal","amount":300.0,"accountNumber":3}]}],"mortgage":{"loanAmount":500.0,"annualRate":0.05,"termYears":10.0,"remainingAmount":200.0}}]
+[{"username":"testing","bankAccounts":[{"balance":950.0,"accountNumber":1,"transactionHistory":[{"type":"Deposit","amount":999.0,"accountNumber":1},{"type":"Withdrawal","amount":9.0,"accountNumber":1},{"type":"Transfer Out","amount":90.0,"accountNumber":2},{"type":"Transfer In","amount":100.0,"accountNumber":4},{"type":"Withdrawal","amount":50.0,"accountNumber":1}]},{"balance":0.0,"accountNumber":2,"transactionHistory":[{"type":"Transfer In","amount":90.0,"accountNumber":1},{"type":"Withdrawal","amount":90.0,"accountNumber":2}]},{"balance":89900.0,"accountNumber":4,"transactionHistory":[{"type":"Deposit","amount":90000.0,"accountNumber":4},{"type":"Transfer Out","amount":100.0,"accountNumber":1}]}],"creditScore":0},{"username":"newmax","bankAccounts":[],"mortgage":{"loanAmount":100.0,"annualRate":0.05,"termYears":10.0,"remainingAmount":10.0},"creditScore":0},{"username":"brad","bankAccounts":[],"mortgage":{"loanAmount":100.0,"annualRate":0.06,"termYears":67.0,"remainingAmount":100.0},"creditScore":0},{"username":"max","bankAccounts":[{"balance":0.0,"accountNumber":3,"transactionHistory":[{"type":"Deposit","amount":300.0,"accountNumber":3},{"type":"Withdrawal","amount":300.0,"accountNumber":3}]}],"mortgage":{"loanAmount":500.0,"annualRate":0.05,"termYears":10.0,"remainingAmount":200.0},"creditScore":0}]
